+++ packages/frontend/src/components/common/ResourceView/index.tsx
@@ -1,11 +1,11 @@
 import { Tooltip2 } from '@blueprintjs/popover2';
 import { assertUnreachable } from '@lightdash/common';
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 import ResourceActionHandlers, {
     ResourceViewItemAction,
     ResourceViewItemActionState,
 } from './ResourceActionHandlers';
-import { ResourceViewItem } from './resourceTypeUtils';
+import { ResourceViewItem, ResourceViewItemType } from './resourceTypeUtils';
 import {
     ResourceEmptyStateWrapper,
     ResourceTag,
@@ -13,11 +13,21 @@ import {
     ResourceViewContainer,
     ResourceViewHeader,
     ResourceViewSpacer,
+    ResourceViewTab,
 } from './ResourceView.styles';
 import ResourceViewGrid from './ResourceViewGrid';
 import ResourceViewList, {
     ResourceViewListCommonProps,
 } from './ResourceViewList';
+
+type Tab = {
+    id: string;
+    name: string;
+    icon?: JSX.Element;
+    sort?: (a: ResourceViewItem, b: ResourceViewItem) => number;
+};
+
+type Group = ResourceViewItemType[];
 
 export interface ResourceViewCommonProps {
     headerTitle?: string;
@@ -25,6 +35,8 @@ export interface ResourceViewCommonProps {
     headerIconTooltipContent?: string;
     headerAction?: React.ReactNode;
     items: ResourceViewItem[];
+    tabs?: Tab[];
+    groups?: Group[];
     showCount?: boolean;
     renderEmptyState?: () => React.ReactNode;
     view?: ResourceViewType;
@@ -40,20 +52,24 @@ type ResourceViewProps = ResourceViewCommonProps & ResourceViewListCommonProps;
 const ResourceView: React.FC<ResourceViewProps> = ({
     view = ResourceViewType.LIST,
     items,
+    tabs,
+    groups,
     headerTitle,
     headerIcon,
     headerIconTooltipContent,
     headerAction,
     enableSorting,
     enableMultiSort,
-    defaultColumnVisibility,
     defaultSort,
+    defaultColumnVisibility,
     showCount = true,
     renderEmptyState,
 }) => {
     const [action, setAction] = useState<ResourceViewItemActionState>({
         type: ResourceViewItemAction.CLOSE,
     });
+
+    const [activeTabId, setActiveTabId] = useState(tabs?.[0]?.id);
 
     const handleAction = useCallback(
         (newAction: ResourceViewItemActionState) => {
@@ -62,10 +78,44 @@ const ResourceView: React.FC<ResourceViewProps> = ({
         [],
     );
 
+    const presortedItems = useMemo(() => {
+        if (!tabs || tabs?.length === 0) return items;
+
+        const activeTab = tabs.find((tab) => tab.id === activeTabId);
+        if (!activeTab || !activeTab.sort) return items;
+
+        return items.sort(activeTab.sort);
+    }, [items, tabs, activeTabId]);
+
+    const sortProps =
+        tabs && tabs?.length > 0
+            ? null
+            : {
+                  enableSorting,
+                  enableMultiSort,
+                  defaultSort,
+              };
+
     return (
         <>
+            {tabs && tabs?.length > 0
+                ? tabs.map((tab) => (
+                      <ResourceViewTab
+                          key={tab.id}
+                          icon={tab.icon}
+                          intent={tab.id === activeTabId ? 'primary' : 'none'}
+                          onClick={() => setActiveTabId(tab.id)}
+                          minimal
+                          selected={activeTabId === tab.id}
+                      >
+                          {tab.name}
+                      </ResourceViewTab>
+                  ))
+                : null}
+
             <ResourceViewContainer>
-                {headerTitle || headerAction ? (
+                {tabs && tabs?.length > 0 ? null : headerTitle ||
+                  headerAction ? (
                     <ResourceViewHeader>
                         {headerTitle && (
                             <ResourceTitle>{headerTitle}</ResourceTitle>
@@ -78,8 +128,10 @@ const ResourceView: React.FC<ResourceViewProps> = ({
                                 {headerIcon}
                             </Tooltip2>
                         )}
-                        {showCount && items.length > 0 && (
-                            <ResourceTag round>{items.length}</ResourceTag>
+                        {showCount && presortedItems.length > 0 && (
+                            <ResourceTag round>
+                                {presortedItems.length}
+                            </ResourceTag>
                         )}
 
                         <ResourceViewSpacer />
@@ -88,7 +140,7 @@ const ResourceView: React.FC<ResourceViewProps> = ({
                     </ResourceViewHeader>
                 ) : null}
 
-                {items.length === 0 ? (
+                {presortedItems.length === 0 ? (
                     !!renderEmptyState ? (
                         <ResourceEmptyStateWrapper>
                             {renderEmptyState()}
@@ -96,15 +148,17 @@ const ResourceView: React.FC<ResourceViewProps> = ({
                     ) : null
                 ) : view === ResourceViewType.LIST ? (
                     <ResourceViewList
-                        items={items}
-                        enableSorting={enableSorting}
-                        enableMultiSort={enableMultiSort}
+                        items={presortedItems}
+                        {...sortProps}
                         defaultColumnVisibility={defaultColumnVisibility}
-                        defaultSort={defaultSort}
                         onAction={handleAction}
                     />
                 ) : view === ResourceViewType.GRID ? (
-                    <ResourceViewGrid items={items} onAction={handleAction} />
+                    <ResourceViewGrid
+                        items={presortedItems}
+                        groups={groups}
+                        onAction={handleAction}
+                    />
                 ) : (
                     assertUnreachable(view, 'Unknown resource view type')
                 )}